Trying to get the code to work with PsychoPy v3.0.5
This newer version using an updated version of pyo. The code that presents sound files seems to be more stable with this update then with the 1.90.1 version of Psychopy.
The window type needs to be changed to pyglet for the code to work.
I am also using the wx code that is embedded in "gui" from psychopy to make sure the GUI rendering is stable.