Finished off the fixed lists for the dots task.
Now the dots task has the option of generating the dot locations on the fly or using a fixed location. The task also runs from a config file for both the behavioral and the fMRI experiments.
The generate lists script is the code to generate the code for the fixed locations.